+++ src/models/rnn_model.cpp
@@ -6,8 +6,8 @@ class RNNModel : public BasicModel {
 public:
 	void Initialize(const nlohmann::json &jConf) override {
 		ArgMan argMan;
-		Arg<uint32_t> argCharsetSize("charset_size", argMan);
-		Arg<uint32_t> argCategories("categories", argMan);
+		Arg<uint64_t> argCharsetSize("charset_size", argMan);
+		Arg<uint64_t> argCategories("categories", argMan);
 		ParseArgsFromJson(jConf, argMan);
 
 		m_i2h = register_module("i2h", torch::nn::Linear(
@@ -19,10 +19,10 @@ public:
 	torch::Tensor Forward(std::vector<torch::Tensor> inputs) override {
 		auto input = inputs[0];
 		CHECK_EQ(input.dim(), 3);
-		size_t nStrLen = input.size(1);
+		uint64_t nStrLen = input.size(1);
 		input = input.transpose(0, 1);
 		auto hidden = torch::zeros_like(input[0]);
-		for (size_t i = 0; i < nStrLen; ++i) {
+		for (uint64_t i = 0; i < nStrLen; ++i) {
 			hidden = torch::cat({input[i], hidden}, 1);
 			if (i < nStrLen - 1) {
 				hidden = m_i2h(hidden);
